// web/account/v1/account_pb.d.ts
export declare enum RegionalRoutingValues {
  /**
   * @generated from enum value: EUROPE = 0;
   */
  EUROPE = 0,

  /**
   * @generated from enum value: AMERICAS = 1;
   */
  AMERICAS = 1,

  /**
   * @generated from enum value: ASIA = 2;
   */
  ASIA = 2,

  /**
   * @generated from enum value: ESPORTS = 3;
   */
  ESPORTS = 3,
}